snappy: unisys: latest code with suggestions from irc
This patch includes my current workign directory. The Snappy package
at this moment is not functional. However, with the help of jdstrand
in irc I able to succesfully create test.out but file is empty. Also
go server is not working anymore must have been a change I made to
the snapcraft.yaml